Escribi cada tipo de Tarjeta en un archivo direfente y volvi el descuento de viajes plus una funcion de la Tarjeta
