+++ src/FranquiciaCompleta.php
@@ -1,0 +1,6 @@
+<?php
+namespace TrabajoTarjeta;
+
+class FranquiciaCompleta extends Tarjeta {
+	public $precio=0.0;
+}
