+++ src/MedioBoleto.php
@@ -1,0 +1,7 @@
+<?php
+namespace TrabajoTarjeta;
+
+class MedioBoleto extends Tarjeta {
+	public $precio=7.40;
+}
+
